Make `AutoconfigUrlProvider` an interface
Have different implementations for an autoconfig server hosted by the email provider (`ProviderAutoconfigUrlProvider`) and the ISP DB (`IspDbAutoconfigUrlProvider`).
