+++ feature/autodiscovery/autoconfig/src/test/kotlin/app/k9mail/autodiscovery/autoconfig/ProviderAutoconfigUrlProviderTest.kt
@@ -1,0 +1,20 @@
+package app.k9mail.autodiscovery.autoconfig
+
+import assertk.assertThat
+import assertk.assertions.containsExactly
+import org.junit.Test
+
+class ProviderAutoconfigUrlProviderTest {
+    private val urlProvider = ProviderAutoconfigUrlProvider()
+
+    @Test
+    fun `getAutoconfigUrls with ASCII email address`() {
+        val autoconfigUrls = urlProvider.getAutoconfigUrls("test@domain.example")
+
+        assertThat(autoconfigUrls.map { it.toString() }).containsExactly(
+            "https://autoconfig.domain.example/mail/config-v1.1.xml?emailaddress=test%40domain.example",
+            "https://domain.example/.well-known/autoconfig/mail/config-v1.1.xml",
+            "http://domain.example/.well-known/autoconfig/mail/config-v1.1.xml",
+        )
+    }
+}
